requirements.txt
kafka-python
apache-pyspark
minio
airflow
